LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_anchor_download_block_downloads.tentative-expected.txt

CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing
CONSOLE MESSAGE: Not allowed to download due to sandboxing


PASS <a download> triggered download in sandbox is blocked.
PASS <a download> triggered download in sandbox is blocked before a request is made.
PASS <a > triggered download in sandbox is blocked.
PASS <a target="_blank" > triggered download in sandbox is blocked.
PASS <a target="_blank" rel="noopener" > triggered download in sandbox is blocked.
